[MACOS] fix sign conversion

+++ macos_hello_world.c
@@ -86,7 +86,7 @@ static long sys_write(unsigned int fd, const char *buf, unsigned long len)
 {
     register long x0 __asm__("x0") = fd;
     register const char *x1 __asm__("x1") = buf;
-    register long x2 __asm__("x2") = len;
+    register long x2 __asm__("x2") = (long)len;
     register long x16 __asm__("x16") = 0x2000004; /* SYS_write */
     __asm__ volatile(
         "svc 0"
